pyproject.toml
[tool.poetry]
name = "cits3403-project"
version = "0.1.0"
description = ""
authors = ["tim <timothy.tan.93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
idna = "^3.3"
importlib-metadata = "^4.11.3"
Jinja2 = "^3.1.1"
Flask-WTF = "^1.0.1"
Flask = "^2.1.1"
pyparsing = "^3.0.7"
python-dateutil = "^2.8.2"
Markdown = "^3.3.6"
MarkupSafe = "^2.1.1"
PyYAML = "^6.0"
pyyaml_env_tag = "^0.1"
requests = "^2.27.1"
six = "^1.16.0"
typing-extensions = "^4.1.1"
watchdog = "^2.1.7"
zipp = "^3.8.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
Flask-Login = "^0.6.0"
email-validator = "^1.1.3"
black = "^22.3.0"
pre-commit = "^2.18.1"
click = "8.0.2"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
dotsi = "^0.0.3"
Flask-Cors = "^3.0.10"
pytest = "^7.1.2"
Authlib = "^1.0.1"

[tool.poetry.dev-dependencies]

[tool.black]
force-exclude = "^migrations/"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
